requirements.txt
streamlit
groq
pandas
psycopg2-binary
selenium
beautifulsoup4
webdriver-manager